evolution data

+++ src/main/java/com/imetro/domain/dto/stats/Stats.java
@@ -3,6 +3,6 @@ package com.imetro.domain.dto.stats;
 import java.time.LocalDateTime;
 import java.util.UUID;
 
-public record  Stats(float velocidade, float precisao, float consistencia, float logica, float resiliencia) {
+public record  Stats(double velocidade, double precisao, double consistencia, double logica, double resiliencia) {
 
 }
